+++ exporter/opentelemetry-exporter-richconsole/tests/test_rich_exporter.py
@@ -96,3 +96,13 @@ def test_multiple_traces(tracer_provider):
         parent_2.name in child.label
         for child in trees[traceid_1].children[0].children
     )
+
+def test_no_deadlock(tracer_provider):
+    # non-regression test for https://github.com/open-telemetry/opentelemetry-python-contrib/issues/3254
+
+    tracer = tracer_provider.get_tracer(__name__)
+    with tracer.start_as_current_span("parent"):
+        with tracer.start_as_current_span("child") as child:
+            pass
+
+    RichConsoleSpanExporter.spans_to_tree((child,))